backend/requirements.txt
flask
flask-cors
python-dotenv
pinecone-client
langchain
langchain_pinecone
langchain_google_genai
huggingface_hub
